creates .__ftr_alternates.text section
creates .__ftr_alternates.text section in vmlinux
and emits relocations using --emit-relocs to retain
relocations needed for feature fixup in the future.

--emit-relocs creates .rela sections in vmlinux which
can later be used for feature fixups.

Without that, just .__ftr_alternates.text section is created.

The __ftr_alt* sections can then be marked executable,
enabling the linker to emit a stub. This removes the restriction
on the placement of __ftr_alt* sections close to .text.

Signed-off-by:Sathvika Vasireddy <sv@linux.ibm.com>